+++ api/src/main/java/com/fitbur/testify/Real.java
@@ -38,17 +38,16 @@ public @interface Real {
     /**
      * <p>
      * Indicates whether mock that delegates to the real instance should be
-     * created. This is usefully if you wish to stub certain method call of the
-     * real instance and/or verify calls to protected/private methods of the
-     * real instance. Be very careful how you use this since delegating to the
-     * real instance can have undesirable side effect.
+     * created. This is useful if you wish to stub or verify c package private
+     * methods of the class under test instance. Be very careful how you use
+     * this since delegating to the real instance can have undesirable side
+     * effect.
      * </p>
-     *
      * <p>
      * By default a delegating mock is not created.
      * </p>
      *
-     * @return true if a mock that delegates to real instance should be created.
+     * @return true if a mock delegating mock is created, false otherwise.
      */
     boolean value() default false;
 }