+++ level/integration/src/main/java/com/fitbur/testify/integration/IntegrationTestVerifier.java
@@ -20,9 +20,14 @@ import com.fitbur.testify.TestVerifier;
 import com.fitbur.testify.descriptor.CutDescriptor;
 import com.fitbur.testify.descriptor.FieldDescriptor;
 import com.fitbur.testify.descriptor.ParameterDescriptor;
+import com.fitbur.testify.need.Need;
 import static com.google.common.base.Preconditions.checkState;
 import static java.lang.reflect.Modifier.isFinal;
+import java.security.AccessController;
+import java.security.PrivilegedAction;
 import java.util.Collection;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.Optional;
 import org.slf4j.Logger;
 
@@ -46,68 +51,116 @@ public class IntegrationTestVerifier implements TestVerifier {
 
     @Override
     public void dependency() {
-        String mockito = "org.mockito.Mockito";
+        AccessController.doPrivileged((PrivilegedAction<Object>) () -> {
+            Map<String, String> dependencies = new HashMap<String, String>() {
+                {
+                    put("org.mockito.Mockito", "Mockito");
+                    put("org.springframework.context.annotation.AnnotationConfigApplicationContext", "Spring Context");
+                }
+            };
 
-        try {
-            Class.forName(mockito);
-        } catch (ClassNotFoundException ex) {
-            checkState(false,
-                    "'%s' not found in the classpath. Please add mockito to the "
-                    + "classpath.",
-                    mockito);
-        }
+            dependencies.entrySet().parallelStream().forEach(p -> {
+                try {
+                    Class.forName(p.getKey());
+                } catch (ClassNotFoundException e) {
+                    checkState(false,
+                            "'%s' not found. Please insure '%s' is in the classpath.",
+                            p.getKey(), p.getValue());
+                }
+            });
+            return null;
+        });
+
     }
 
     @Override
     public void configuration() {
-        String testClassName = context.getTestClassName();
-        CutDescriptor cutDescriptor = context.getCutDescriptor();
-        Collection<FieldDescriptor> fieldDescriptors = context.getFieldDescriptors().values();
+        AccessController.doPrivileged((PrivilegedAction<Object>) () -> {
+            String testClassName = context.getTestClassName();
+            CutDescriptor cutDescriptor = context.getCutDescriptor();
+            Collection<FieldDescriptor> fieldDescriptors = context.getFieldDescriptors().values();
 
-        if (cutDescriptor == null) {
-            checkState(!fieldDescriptors.isEmpty(),
-                    "Test class '%s' does not define a field annotated with @Cut "
-                    + "nor does it have fields annotated with @Real or @Inject. "
-                    + "Please insure the test class declares one field annotated "
-                    + "with @Cut or declares fields annotated with @Real or @Real.",
-                    testClassName
-            );
-        }
+            if (context.getCutCount() == 1) {
+                checkState(context.getConstructorCount() == 1,
+                        "Class under test '%s' has '%d' constructors. Please insure that "
+                        + "the class under test has one and only one constructor.",
+                        cutDescriptor.getTypeName(), context.getConstructorCount());
+            }
+            // insure that only one field has Cut annotation on it.
+            if (context.getCutCount() > 1) {
+                checkState(false,
+                        "Found more than one class under test in %s. Please insure "
+                        + "that only one field is annotated with @Cut.",
+                        testClassName);
+            }
 
-        fieldDescriptors.parallelStream().forEach(p -> {
-            Class<?> fieldType = p.getType();
-            String fieldName = p.getName();
-            String fieldTypeName = p.getTypeName();
+            //insure that there is a field annotated with @Cut defined or one or more
+            //fields annotated with @Real or @Inject
+            if (context.getCutCount() == 0 && fieldDescriptors.isEmpty()) {
+                checkState(false,
+                        "Test class '%s' does not define a field annotated with @Cut "
+                        + "nor does it have fields annotated with @Real or @Inject. "
+                        + "Please insure the test class declares a single field annotated "
+                        + "with @Cut or declares fields annotated with @Real or @Real.",
+                        testClassName
+                );
+            }
 
-            checkState(!isFinal(fieldType.getModifiers()),
-                    "Field '%s' in test class '%s' can not be mocked because '%s'"
-                    + " is a final class.",
-                    fieldName, testClassName, fieldTypeName);
+            //insure need providers have default constructors.
+            context.getNeeds().parallelStream().map(Need::value).forEach(p -> {
+                try {
+                    p.getDeclaredConstructor();
+                } catch (NoSuchMethodException e) {
+                    checkState(false,
+                            "Need provider '%s' defined in test class '%s' does not have a "
+                            + "zero argument default constructor. Please insure that the need "
+                            + "provider defines accessible zero argument default constructor.",
+                            testClassName, p.getSimpleName()
+                    );
+                }
+            });
 
+            fieldDescriptors.parallelStream().forEach(p -> {
+                Class<?> fieldType = p.getType();
+                String fieldName = p.getName();
+                String fieldTypeName = p.getTypeName();
+
+                checkState(!isFinal(fieldType.getModifiers()),
+                        "Field '%s' in test class '%s' can not be mocked because '%s'"
+                        + " is a final class.",
+                        fieldName, testClassName, fieldTypeName);
+
+            });
+
+            return null;
         });
     }
 
     @Override
     public void wiring() {
-        CutDescriptor cutDescriptor = context.getCutDescriptor();
-        String testClassName = context.getTestClassName();
-        Collection<ParameterDescriptor> paramDescriptors
-                = context.getParamaterDescriptors().values();
+        AccessController.doPrivileged((PrivilegedAction<Object>) () -> {
+            CutDescriptor cutDescriptor = context.getCutDescriptor();
+            String testClassName = context.getTestClassName();
+            Collection<ParameterDescriptor> paramDescriptors
+                    = context.getParamaterDescriptors().values();
 
-        if (cutDescriptor != null) {
-            String cutClassName = cutDescriptor.getTypeName();
-            paramDescriptors.parallelStream().forEach(p -> {
-                Optional instance = p.getInstance();
-                if (!instance.isPresent()) {
-                    String paramTypeName = p.getTypeName();
-                    logger.warn("Improper wiring in detected. Class under test '{}' defined "
-                            + "in '{}' declars constructor argument of type '{}' but '{}' "
-                            + "does not define a field of type '{}' annotated with @Mock.",
-                            cutClassName, testClassName, paramTypeName, testClassName, paramTypeName);
-                }
+            if (cutDescriptor != null) {
+                String cutClassName = cutDescriptor.getTypeName();
+                paramDescriptors.parallelStream().forEach(p -> {
+                    Optional instance = p.getInstance();
+                    if (!instance.isPresent()) {
+                        String paramTypeName = p.getTypeName();
+                        logger.warn("Improper wiring detected. Class under test '{}' defined "
+                                + "in '{}' declars constructor argument of type '{}' but '{}' "
+                                + "does not define a field of type '{}' annotated with @Mock.",
+                                cutClassName, testClassName, paramTypeName, testClassName, paramTypeName);
+                    }
 
-            });
-        }
+                });
+            }
+
+            return null;
+        });
     }
 
 }